backend/requirements.txt
# HireQ - Backend Dependencies
# Python 3.10+

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.1
beanie==1.25.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
PyJWT==2.11.0

# File Processing
python-docx==1.1.0
PyMuPDF>=1.24.0
pdfplumber==0.10.3

# NLP & ML
spacy==3.7.2
sentence-transformers==2.2.2
scikit-learn==1.4.0
numpy==1.26.3
torch==2.1.2

# Sentiment Analysis
transformers==4.36.2
textblob==0.17.1

# Speech-to-Text (Local Whisper - free, no API key needed)
openai-whisper==20231117
ffmpeg-python==0.2.0

# File Storage
boto3==1.34.14
python-dotenv==1.0.0

# PDF Generation
reportlab==4.0.8
weasyprint==60.2

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1
aiofiles==23.2.1

# AI Chatbot
google-genai>=1.0.0
chromadb>=0.5.0

# CORS
starlette==0.35.1

# Email
resend>=0.7.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

